Refactor file methods into NodeFileUtils
Node packages like fs and path with native dependencies are not available on the browser,
and reading files on the browser will involve making a HTTP request to fetch the JSON file.

To avoid invoking such methods in the browser, they have been moved into
NodeFileUtils in the build-time folder, with the prefix "Node" for clarity.